Session Implementacao

+++ src/main/java/com/br/resistenem/controller/AlternativaController.java
@@ -1,5 +1,7 @@
 package com.br.resistenem.controller;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -18,7 +20,10 @@ public class AlternativaController {
 
 
 	@RequestMapping(value="/alternativa/insertAlternativa", method=RequestMethod.POST)
-	public String insertAlternativa(Alternativa alternativa, RedirectAttributes attibutes) {
+	public String insertAlternativa(Alternativa alternativa, RedirectAttributes attibutes, HttpSession session) {
+		if ("false".equals(session.getAttribute("isLogado").toString())) {
+			return "redirect:/";
+		}
 		if ("".equals(alternativa.getAlternativa()) || "".equals(alternativa.getIdQuestao())) {
 			attibutes.addFlashAttribute("menssagem", "verifique os campos!");
 			attibutes.addFlashAttribute("error", true);
@@ -31,7 +36,10 @@ public class AlternativaController {
 	}
 	
 	@RequestMapping(value="/alternativa/updateAlternativa", method=RequestMethod.POST)
-	public String updateAlternativa(Alternativa alternativa, RedirectAttributes attibutes) {
+	public String updateAlternativa(Alternativa alternativa, RedirectAttributes attibutes, HttpSession session) {
+		if ("false".equals(session.getAttribute("isLogado").toString())) {
+			return "redirect:/";
+		}
 		if ("".equals(alternativa.getAlternativa()) || "".equals(alternativa.getIdQuestao())) {
 			attibutes.addFlashAttribute("menssagem", "verifique os campos!");
 			attibutes.addFlashAttribute("error", true);
@@ -44,13 +52,19 @@ public class AlternativaController {
 	}
 	
 	@RequestMapping(value="/alternativa/editarAlternativa/{id}", method=RequestMethod.GET)
-	public String editarAlternativa(@PathVariable("id") String id) {
+	public String editarAlternativa(@PathVariable("id") String id, HttpSession session) {
+		if ("false".equals(session.getAttribute("isLogado").toString())) {
+			return "redirect:/";
+		}
 		return "redirect:/questao/editarAlternativa/"+id;
 
 	}
 	
 	@RequestMapping("/alternativa/excluirAlternativa/{id}")
-	public String excluirAlternativa(@PathVariable("id") String id, RedirectAttributes attibutes) {
+	public String excluirAlternativa(@PathVariable("id") String id, RedirectAttributes attibutes, HttpSession session) {
+		if ("false".equals(session.getAttribute("isLogado").toString())) {
+			return "redirect:/";
+		}
 		Alternativa alternativa = ar.findAllById(id);
 		ar.deleteById(id);
 		attibutes.addFlashAttribute("menssagem", "Alternativa excluida com sucesso!");
